Fix missed refactors, adjustments
Table.php didn't move the InitialTablePart parameter for one call
Fix using IN in a search term if a CSV string was passed instead of an array
If no search term was specified in a badge search, don't blow up the rest of the possible searchTerms that will be generated
Also, expose the AdjustSearchTerms utility function as it's useful in general, may refactor out to the Table class?
shop.js check if errorCb is a function before blindly attempting to call it.