More work on locations and assignments
Assignment->search: Provide only the assignments that belong to the event
Location->Assignments: Provide assignments for a particular location ad-hoc
Location->Available applications: Provide applications
Location->CRU: enhance to include assignments associated with that location
Public APIs to get the listed categories/locations
Assignments must have categories for the events list
Short code for locations should be unique per event
Expose the Notes field to the badge listings
Expose approved assignment count and actual assigned count, as well as a simple list of their IDs
Fix v-model reactivity in a few components. If it truly has no side effects, apply to the rest of them that use skipEmitOnce
Add retrieval and caching of assignments and categories. Cache invalidation isn't yet implemented because a few cascading things need to be done in the backend