Make events in grid draggable
Grid tab can be toggled to allow edits. Events can be easily dragged between locations and days as you might expect. Might possibly add a double-click handler to pop the advanced editor.
Made the locations dropdown in the grid sorted and include the short code for searches
More appropriately prevent user selecting text on events. It doesn't prevent text from outside the time slots being selected in some cases, but it's good enough to not be distracting most of the time.