Badge format and preprinting work
Allow the pre-printing search to specify only not-printed ones, and include unprintable (i.e. unpaid) in the list for selection
Don't individually add badges to the queue in step 2, will be done in step 3 when the command to start printing is given
A lot of experimental garbage CSS to help fix some issues with badge printing and printing in general.
Make the badge format editor's preview of the printed badge work like the actual prints.
Adjustments to BadgeSearchList component and friends to allow exposing the Select tickbox props. Also emit the resultant retrieved data
Log a bit more data if the template engine doesn't have a field that was requested. Probably should find a way to bubble that as an error...